Moved task item into its own component
